refactor the identifier generation from molecules into own module
split up the parser module into a generator module and move all
related functions.
